+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Add delete option to post 3-dot menu
+
 ## v0.9.18 - 2023-06-27
 - Change post's vote count loading animation to be circular
 - Added post style settings pane